tetragon: Move generic_process_filter to generic functions
Moving generic_process_filter to generic functions and changing it to
use maps directly.

Passing map pointer through argument might end up storing and loading
it to/from heap which makes it scalar value and breaks verifier.
