+++ bpf/process/bpf_generic_kprobe.c
@@ -12,7 +12,6 @@
 #include "retprobe_map.h"
 #include "types/operations.h"
 #include "types/basic.h"
-#include "pfilter.h"
 #include "policy_filter.h"
 
 char _license[] __attribute__((section("license"), used)) = "Dual BSD/GPL";
@@ -97,8 +96,7 @@ generic_kprobe_process_filter(void *ctx)
 {
 	int ret;
 
-	ret = generic_process_filter((struct bpf_map_def *)&process_call_heap,
-				     (struct bpf_map_def *)&filter_map);
+	ret = generic_process_filter();
 	if (ret == PFILTER_CONTINUE)
 		tail_call(ctx, &kprobe_calls, TAIL_CALL_FILTER);
 	else if (ret == PFILTER_ACCEPT)
