+++ bpf/process/pfilter.h
@@ -2,8 +2,7 @@
 #define __PFILTER_H__
 
 /**
- * Process filters
- * see generic_process_filter below
+ * Process filters (see generic_process_filter)
  */
 
 #define FIND_PIDSET(value, isns)                                          \
@@ -540,72 +539,4 @@ process_filter_done(struct msg_selector_data *sel,
 		return PFILTER_ACCEPT;
 	return PFILTER_REJECT;
 }
-
-// generic_process_filter performs first pass filtering based on pid/nspid.
-// We keep a list of selectors that pass.
-//
-// if filter check was successful, it will return PFILTER_ACCEPT and properly
-// set the values of:
-//    current->pid
-//    current->ktime
-// for the memory located at index 0 of @msg_heap assuming the value follows the
-// msg_generic_hdr structure.
-FUNC_INLINE int
-generic_process_filter(struct bpf_map_def *heap, struct bpf_map_def *fmap)
-{
-	struct execve_map_value *enter;
-	struct msg_generic_kprobe *msg;
-	struct msg_execve_key *current;
-	struct msg_selector_data *sel;
-	int curr, zero = 0;
-	bool walker = 0;
-	__u32 ppid;
-
-	msg = map_lookup_elem(heap, &zero);
-	if (!msg)
-		return 0;
-
-	enter = event_find_curr(&ppid, &walker);
-	if (enter) {
-		int selectors, pass;
-		__u32 *f = map_lookup_elem(fmap, &msg->idx);
-
-		if (!f)
-			return PFILTER_ERROR;
-
-		sel = &msg->sel;
-		current = &msg->current;
-
-		curr = sel->curr;
-		if (curr > MAX_SELECTORS)
-			return process_filter_done(sel, enter, current);
-
-		selectors = f[0];
-		/* If no selectors accept process */
-		if (!selectors) {
-			sel->pass = true;
-			return process_filter_done(sel, enter, current);
-		}
-
-		/* If we get here with reference to uninitialized selector drop */
-		if (selectors <= curr)
-			return process_filter_done(sel, enter, current);
-
-		pass = selector_process_filter(f, curr, enter, msg);
-		if (pass) {
-			/* Verify lost that msg is not null here so recheck */
-			asm volatile("%[curr] &= 0x1f;\n"
-				     : [curr] "+r"(curr));
-			sel->active[curr] = true;
-			sel->active[SELECTORS_ACTIVE] = true;
-			sel->pass |= true;
-		}
-		sel->curr++;
-		if (sel->curr > selectors)
-			return process_filter_done(sel, enter, current);
-		return PFILTER_CONTINUE; /* will iterate to the next selector */
-	}
-	return PFILTER_CURR_NOT_FOUND;
-}
-
 #endif /* __PFILTER_H__ */
